notification button done with css
