+++ src/components/Header/index.css
@@ -69,12 +69,24 @@
 .Header-notification-body{
     width:100%;
     margin-top:10px;
-    min-width: 100px;
+    min-width: 200px;
     border-radius: 5px;
     position:absolute;
-    left:-45%;
+    transform: translateX(-45%);
     background: black;
     align-self: center;
+}
+.Header-notification-number{
+    width: 15px;
+    height: 15px;
+    text-align: center;
+    background: red;
+    padding: 2px;
+    position: absolute;
+    border-radius: 100%;
+    top: -5px;
+    right: 45px;
+    padding-top: 0px;
 }
 .Header-logout-link{
     width:60%;
@@ -90,6 +102,7 @@
 .Header-notification{
     padding-right: 50px;
     margin-top: 15px;
+    position: relative;
 }
 .Header-user{
     position: relative;
